pyproject.toml
[project]
name = "valorant-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipykernel>=7.2.0",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
]
